+++ plantcv/plantcv/filters/__init__.py
@@ -1,5 +1,5 @@
 from plantcv.plantcv.filters.eccentricity import eccentricity
-from plantcv.plantcv.filters.filter_objs import obj_props
+from plantcv.plantcv.filters.obj_props import obj_props
 
 __all__ = ["eccentricity",
            "obj_props"]
