+++ plantcv/plantcv/filters/obj_props.py
